Move each plugins to a file
