+++ api/lib/infrastructure/plugins/i18n.js
@@ -1,0 +1,12 @@
+module.exports = {
+  plugin: require('hapi-i18n'),
+  options: {
+    locales: ['en', 'fr'],
+    directory: __dirname + '/../../../translations',
+    defaultLocale: 'fr',
+    queryParameter: 'lang',
+    languageHeaderField: 'Accept-Language',
+    objectNotation: true,
+    updateFiles: false,
+  },
+};
